+++ platform/jvm/capture/src/main/kotlin/io/bitdrift/capture/IBridge.kt
@@ -40,7 +40,6 @@ internal interface IBridge {
         errorReporter: IErrorReporter,
         startInSleepMode: Boolean,
         issueCallbackConfiguration: IssueCallbackConfiguration?,
-        startupReplayEligibility: Int,
         initialFields: Array<Field> = emptyArray(),
     ): Long
 }